Refactor: move filter function to binary
